requirements.txt
Flask
gunicorn
requests
Pillow
